requirements-streamlit.txt
streamlit==1.45.1
altair>=5.0.0
pandas==2.2.2
psycopg2-binary==2.9.9
PyYAML>=6.0.1
